requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
databases[postgresql]
python-jose
passlib[bcrypt]
python-multipart
jinja2
aiofiles